backend/requirements/local.txt
-r base/base.txt

validation==0.8.1
drf-yasg==1.20.0
model_bakery==1.5.0
pytest-django==4.5.2
pytest-xdist==2.5.0
pytest-benchmark==3.4.1
pytest-cov==3.0.0